requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.29
Flask-SQLAlchemy==3.1.1
lambda-packages
troposphere
zappa
flask-cors==6.0.0
psycopg2-binary
pytest==8.3.5
pytest-flask
pytest-cov
pytest-order==1.3.0
mock
pylint
flask-mail
random-password-generator
PyJWT==2.9.0
pyjwt[crypto]
cryptography==44.0.1
boto3==1.39.10
openai==1.12.0
redis==5.1.1
pydantic==2.9.2
strands-agents==1.1.0